requirements.txt
py==1.4.31
pytest==2.9.1
python-Levenshtein==0.12.0
